requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=py311h06a4308_0
accelerate=0.33.0=pypi_0
aiohttp=3.9.5=py311h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
albucore=0.0.13=pypi_0
albumentations=1.4.13=pypi_0
annotated-types=0.7.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=py311h06a4308_0
blas=1.0=mkl
blinker=1.6.2=py311h06a4308_0
bpy=4.2.0=pypi_0
brotli-python=1.0.9=py311h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.7.2=h06a4308_0
cachetools=5.3.3=py311h06a4308_0
certifi=2024.7.4=py311h06a4308_0
cffi=1.16.0=py311h5eee18b_1
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.7=py311h06a4308_0
clip=1.0=pypi_0
colour=0.1.5=pypi_0
comm=0.2.1=py311h06a4308_0
contourpy=1.2.1=pypi_0
cryptography=42.0.5=py311hdda0065_1
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.127=h6a678d5_0
cuda-runtime=12.1.0=0
cuda-version=12.4=hbda6634_3
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
debugpy=1.6.7=py311h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
eval-type-backport=0.2.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.13.1=py311h06a4308_0
fonttools=4.51.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py311h5eee18b_0
fsspec=2024.3.1=pypi_0
ftfy=6.2.0=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
google-auth=2.29.0=py311h06a4308_0
google-auth-oauthlib=0.5.2=py311h06a4308_0
grpcio=1.62.2=py311h6a678d5_0
huggingface-hub=0.24.2=pypi_0
idna=3.7=py311h06a4308_0
imageio=2.34.0=pypi_0
imath=0.0.2=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.28.0=py311h06a4308_0
ipython=8.23.0=pypi_0
jedi=0.19.1=py311h06a4308_0
jinja2=3.1.4=py311h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_2
jupyter_client=8.6.0=py311h06a4308_0
jupyter_core=5.7.2=py311h06a4308_0
kiwisolver=1.4.5=pypi_0
lazy-loader=0.4=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.1.3=h99ab3db_1
libcurand=10.3.5.147=h99ab3db_1
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libgrpc=1.62.2=h2d74bed_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libprotobuf=4.25.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lightning-utilities=0.11.6=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.4.1=py311h06a4308_0
markupsafe=2.1.3=py311h5eee18b_0
mathutils=3.3.0=pypi_0
matplotlib=3.8.4=pypi_0
matplotlib-inline=0.1.6=py311h06a4308_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multidict=6.0.4=py311h5eee18b_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=py311h06a4308_0
networkx=3.3=py311h06a4308_0
numpy=1.26.4=py311h08b1b3b_0
numpy-base=1.26.4=py311hf175353_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.5.82=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=py311h06a4308_0
ocl-icd=2.3.2=h5eee18b_1
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openexr=3.2.4=pypi_0
openjpeg=2.4.0=h9ca470c_2
openssl=3.0.14=h5eee18b_0
packaging=24.1=py311h06a4308_0
pandas=2.2.2=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pillow=10.4.0=py311h5eee18b_0
pip=24.1.2=pypi_0
platformdirs=3.10.0=py311h06a4308_0
prompt-toolkit=3.0.43=py311h06a4308_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=4.25.3=py311h12ddb61_0
psutil=5.9.0=py311h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycocotools=2.0.8=pypi_0
pycparser=2.22=pypi_0
pydantic=2.8.2=pypi_0
pydantic-core=2.20.1=pypi_0
pygments=2.15.1=py311h06a4308_1
pyjwt=2.8.0=py311h06a4308_0
pyopenssl=24.0.0=py311h06a4308_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0post0=py311h06a4308_2
pytorch-cuda=12.1=ha16c6d3_5
pytorch-lightning=2.3.3=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
pyyaml=6.0.1=py311h5eee18b_0
pyzmq=25.1.2=py311h6a678d5_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.32.3=py311h06a4308_0
requests-oauthlib=2.0.0=py311h06a4308_0
rsa=4.7.2=pyhd3eb1b0_1
safetensors=0.4.3=pypi_0
scikit-image=0.23.1=pypi_0
scikit-learn=1.6.0=pypi_0
scipy=1.13.0=pypi_0
setuptools=69.5.1=py311h06a4308_0
shortuuid=1.0.13=pypi_0
six=1.16.0=pyhd3eb1b0_1
soundfile=0.12.1=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.12=py311h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.12.1=py311h06a4308_0
tensorboard-data-server=0.7.0=py311h52d8a92_0
tensorboard-plugin-wit=1.6.0=py_0
tensorboardx=2.6.2.2=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.2.12=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.0.1=pypi_0
torch=2.4.0=pypi_0
torch-tb-profiler=0.4.3=pypi_0
torchaudio=2.2.2=py311_cu121
torchmetrics=1.4.0.post0=pypi_0
torchvision=0.15.2=cpu_py311h6e929fa_0
tornado=6.4.1=py311h5eee18b_0
tqdm=4.66.2=pypi_0
traitlets=5.14.3=py311h06a4308_0
triton=3.0.0=pypi_0
typing_extensions=4.11.0=py311h06a4308_0
tzdata=2024.1=pypi_0
unreal=0.1.1=pypi_0
unrealcv=1.1.0=pypi_0
urllib3=2.2.2=py311h06a4308_0
wcwidth=0.2.13=pypi_0
werkzeug=3.0.3=py311h06a4308_0
wheel=0.43.0=py311h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.3=py311h5eee18b_0
zeromq=4.3.5=h6a678d5_0
zlib=1.2.13=h5eee18b_1
zstandard=0.23.0=pypi_0
zstd=1.5.5=hc292b87_2
